chore: reorganize project structure before vite migration
