+++ src/App.js
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
-import Header from './components/Header/Header';
-import Footer from './components/Footer/Footer';
+import Header from './components/common/Header';
+import Footer from './components/common/Footer';
 import Main from './pages/Main';
 import Login from './pages/Login';
 import Profile from './pages/Profile';
