+++ src/pages/Login.jsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { loginSuccess, loginFailure } from '../redux/actions/authActions';
 import Kakao from '../components/Login/Kakao';
 import Naver from '../components/Login/Naver';
-import '../styles/Login/Login.scss';
+import '../styles/pages/Login.scss';
 
 export default function Login() {
   const [userId, setUserId] = useState('');
