+++ src/styles/pages/Main.scss
