+++ src/styles/pages/Register.scss
