+++ src/styles/pages/Search.scss
