+++ src/styles/pages/Write.scss
